requirements.txt
pytest==6.0.1
requests==2.24.0
Pillow==7.2.0
